chore: split large modules into focused files
