+++ lib/lutaml/xmi/parsers/xmi_base.rb
@@ -6,6 +6,9 @@ require "xmi"
 require "digest"
 require_relative "xmi_connector"
 require_relative "xmi_diagram"
+require_relative "xmi_generalization"
+require_relative "xmi_enum_data_type"
+require_relative "xmi_class_members"
 
 module Lutaml
   module Xmi
@@ -195,90 +198,6 @@ module Lutaml
           klass_hash
         end
 
-        # @param klass [Lutaml::Model::Serializable]
-        # # @return [Hash]
-        def serialize_generalization(klass, options = {})
-          general_hash, next_general_node_id = get_top_level_general_hash(
-            klass, options
-          )
-          return general_hash unless next_general_node_id
-
-          general_hash[:general] = serialize_generalization_attributes(
-            next_general_node_id, options
-          )
-
-          general_hash
-        end
-
-        # @param klass [Lutaml::Model::Serializable]
-        # @return [Array<Hash>]
-        def get_top_level_general_hash(klass, options = {}) # rubocop:disable Metrics/AbcSize,Metrics/MethodLength
-          general_hash, next_general_node_id = get_general_hash(
-            klass.id, options
-          )
-          general_hash[:name] = klass.name
-          general_hash[:type] = klass.type
-          general_hash[:definition] =
-            lookup_element_prop_documentation(klass.id)
-          general_hash[:stereotype] = doc_node_attribute_value(
-            klass.id, "stereotype"
-          )
-
-          [general_hash, next_general_node_id]
-        end
-
-        # @param xmi_id [String]
-        # @param model [Lutaml::Model::Serializable]
-        # @return [Array<Hash>]
-        # @note get generalization node and its owned attributes
-        def serialize_generalization_attributes(general_id, options = {}) # rubocop:disable Metrics/AbcSize,Metrics/MethodLength
-          general_hash, next_general_node_id = get_general_hash(general_id,
-                                                                options)
-
-          if next_general_node_id
-            general_hash[:general] = serialize_generalization_attributes(
-              next_general_node_id, options
-            )
-          end
-
-          general_hash
-        end
-
-        # @param general_node [Lutaml::Model::Serializable]
-        # @return [Hash]
-        def get_general_attributes(general_node)
-          serialize_class_attributes(general_node, with_assoc: true)
-        end
-
-        # @param general_node [Lutaml::Model::Serializable]
-        # @return [String]
-        def get_next_general_node_id(general_node)
-          general_node.generalization.first&.general
-        end
-
-        # @param general_id [String]
-        # @return [Array<Hash>]
-        def get_general_hash(general_id, options = {}) # rubocop:disable Metrics/AbcSize,Metrics/MethodLength
-          general_node = find_packaged_element_by_id(general_id)
-          return [] unless general_node
-
-          general_node_attrs = get_general_attributes(general_node)
-          general_upper_klass = find_upper_level_packaged_element(general_id)
-          next_general_node_id = get_next_general_node_id(general_node)
-
-          [
-            {
-              general_id: general_id,
-              general_name: general_node.name,
-              general_attributes: general_node_attrs,
-              general_upper_klass: ::Lutaml::Xmi::LiquidDrops::PackageDrop
-                .new(general_upper_klass, nil, options),
-              general: {},
-            },
-            next_general_node_id,
-          ]
-        end
-
         # @param id [String]
         # @return [Lutaml::Model::Serializable]
         def find_packaged_element_by_id(id)
@@ -419,199 +338,6 @@ module Lutaml
         # @return [Lutaml::Model::Serializable]
         def find_packaged_element_by_name(name)
           xmi_index.packaged_elements.find { |e| e.name == name }
-        end
-
-        # @param package [Lutaml::Model::Serializable]
-        # @return [Array<Hash>]
-        # @note xpath ./packagedElement[@xmi:type="uml:Enumeration"]
-        def serialize_model_enums(package) # rubocop:disable Metrics/MethodLength
-          enums = package.packaged_element.select do |e|
-            e.type?("uml:Enumeration")
-          end
-
-          enums.map do |enum|
-            {
-              xmi_id: enum.id,
-              name: enum.name,
-              values: serialize_enum_owned_literal(enum),
-              definition: doc_node_attribute_value(enum.id, "documentation"),
-              stereotype: doc_node_attribute_value(enum.id, "stereotype"),
-            }
-          end
-        end
-
-        # @param model [Lutaml::Model::Serializable]
-        # @return [Hash]
-        # @note xpath .//ownedLiteral[@xmi:type="uml:EnumerationLiteral"]
-        def serialize_enum_owned_literal(enum) # rubocop:disable Metrics/MethodLength
-          owned_literals = enum.owned_literal.select do |owned_literal|
-            owned_literal.type? "uml:EnumerationLiteral"
-          end
-
-          owned_literals.map do |owned_literal|
-            # xpath .//type
-            uml_type_id = owned_literal&.uml_type&.idref
-
-            {
-              name: owned_literal.name,
-              type: lookup_entity_name(uml_type_id) || uml_type_id,
-              definition: lookup_attribute_documentation(owned_literal.id),
-            }
-          end
-        end
-
-        # @param model [Lutaml::Model::Serializable]
-        # @return [Array<Hash>]
-        # @note xpath ./packagedElement[@xmi:type="uml:DataType"]
-        def serialize_model_data_types(model) # rubocop:disable Metrics/AbcSize,Metrics/MethodLength
-          all_data_type_elements = []
-          select_all_packaged_elements(
-            all_data_type_elements, model, "uml:DataType"
-          )
-          all_data_type_elements.map do |klass|
-            {
-              xmi_id: klass.id,
-              name: klass.name,
-              attributes: serialize_class_attributes(klass),
-              operations: serialize_class_operations(klass),
-              associations: serialize_model_associations(klass.id),
-              constraints: serialize_class_constraints(klass.id),
-              is_abstract: doc_node_attribute_value(klass.id, "isAbstract"),
-              definition: doc_node_attribute_value(klass.id, "documentation"),
-              stereotype: doc_node_attribute_value(klass.id, "stereotype"),
-            }
-          end
-        end
-
-        # @param klass [Lutaml::Model::Serializable]
-        # @return [Array<Hash>]
-        # @note xpath .//ownedOperation
-        def serialize_class_operations(klass) # rubocop:disable Metrics/MethodLength
-          klass.owned_operation.filter_map do |operation|
-            uml_type = operation.uml_type.first
-            uml_type_idref = uml_type.idref if uml_type
-
-            if operation.association.nil?
-              {
-                id: operation.id,
-                xmi_id: uml_type_idref,
-                name: operation.name,
-                definition: lookup_attribute_documentation(operation.id),
-              }
-            end
-          end
-        end
-
-        # @param klass_id [String]
-        # @return [Array<Hash>]
-        # @note xpath ./constraints/constraint
-        def serialize_class_constraints(klass_id) # rubocop:disable Metrics/MethodLength
-          connector_node = fetch_connector(klass_id)
-
-          if connector_node
-            # In ea-xmi-2.5.1, constraints are moved to source/target under
-            # connectors
-            constraints = %i[source target].map do |st|
-              connector_node.send(st).constraints.constraint
-            end.flatten
-
-            constraints.map do |constraint|
-              {
-                name: HTMLEntities.new.decode(constraint.name),
-                type: constraint.type,
-                weight: constraint.weight,
-                status: constraint.status,
-              }
-            end
-          end
-        end
-
-        # @param klass_id [String]
-        # @return [Lutaml::Model::Serializable]
-        # @note xpath %(//element[@xmi:idref="#{klass['xmi:id']}"])
-        def fetch_element(klass_id)
-          xmi_index.find_element(klass_id)
-        end
-
-        # @param klass [Lutaml::Model::Serializable]
-        # @param with_assoc [Boolean]
-        # @return [Array<Hash>]
-        # @note xpath .//ownedAttribute[@xmi:type="uml:Property"]
-        def serialize_class_attributes(klass, with_assoc: false) # rubocop:disable Metrics/AbcSize,Metrics/CyclomaticComplexity,Metrics/MethodLength
-          owned_attributes = klass.owned_attribute.select do |attr|
-            attr.type?("uml:Property")
-          end
-
-          owned_attributes.filter_map do |oa|
-            if with_assoc || oa.association.nil?
-              attrs = build_class_attributes(oa)
-
-              if with_assoc && oa.association
-                attrs[:association] = oa.association
-                attrs[:definition] = loopup_assoc_def(oa.association)
-                attrs[:type_ns] = get_ns_by_xmi_id(attrs[:xmi_id])
-              end
-
-              attrs
-            end
-          end
-        end
-
-        def loopup_assoc_def(association)
-          connector = fetch_connector(association)
-          connector&.documentation&.value
-        end
-
-        # @param type [String]
-        # @return [String]
-        def get_ns_by_type(type)
-          return unless type
-
-          p = find_klass_packaged_element_by_name(type)
-          return unless p
-
-          find_upper_level_packaged_element(p.id)&.name
-        end
-
-        # @param xmi_id [String]
-        # @return [String]
-        def get_ns_by_xmi_id(xmi_id)
-          return unless xmi_id
-
-          p = find_packaged_element_by_id(xmi_id)
-          return unless p
-
-          find_upper_level_packaged_element(p.id)&.name
-        end
-
-        # @param klass_id [String]
-        # @return [Array<Hash>]
-        def build_class_attributes(owned_attr) # rubocop:disable Metrics/MethodLength
-          uml_type = owned_attr.uml_type
-          uml_type_idref = uml_type.idref if uml_type
-
-          {
-            id: owned_attr.id,
-            name: owned_attr.name,
-            type: lookup_entity_name(uml_type_idref) || uml_type_idref,
-            xmi_id: uml_type_idref,
-            is_derived: owned_attr.is_derived,
-            cardinality: cardinality_min_max_value(
-              owned_attr.lower_value&.value,
-              owned_attr.upper_value&.value,
-            ),
-            definition: lookup_attribute_documentation(owned_attr.id),
-          }
-        end
-
-        # @param min [String]
-        # @param max [String]
-        # @return [Hash]
-        def cardinality_min_max_value(min, max)
-          {
-            min: min,
-            max: max,
-          }
         end
 
         # @node [Lutaml::Model::Serializable]
